[module/std] Split modules for primitive nodes
