+++ include/yave/module/std/bool/bool.hpp
@@ -1,0 +1,14 @@
+//
+// Copyright (c) 2019 mocabe (https://github.com/mocabe)
+// Distributed under LGPLv3 License. See LICENSE for more details.
+//
+
+#pragma once
+
+#include <yave/module/std/primitive/primitive.hpp>
+
+namespace yave {
+
+  YAVE_DECL_PRIM_NODE(Bool, Std.Bool);
+
+}
