+++ include/yave/module/std/num/num.hpp
@@ -1,0 +1,15 @@
+//
+// Copyright (c) 2019 mocabe (https://github.com/mocabe)
+// Distributed under LGPLv3 License. See LICENSE for more details.
+//
+
+#pragma once
+
+#include <yave/module/std/primitive/primitive.hpp>
+
+namespace yave {
+
+  YAVE_DECL_PRIM_NODE(Int, Std.Num);
+  YAVE_DECL_PRIM_NODE(Float, Std.Num);
+
+} // namespace yave
