+++ include/yave/module/std/primitive/primitive.hpp
@@ -9,14 +9,20 @@
 #include <yave/module/std/config.hpp>
 #include <yave/obj/primitive/primitive.hpp>
 
-#define YAVE_DECL_PRIM_NODE(TYPE)                                          \
+#define YAVE_DECL_PRIM_NODE(TYPE, MODULE)                                  \
   namespace node {                                                         \
     using TYPE = PrimitiveCtor<yave::TYPE>;                                \
   }                                                                        \
   template <>                                                              \
   constexpr char node_declaration_traits<node::TYPE>::node_name[] = #TYPE; \
+  template <>                                                              \
+  constexpr char node_declaration_traits<node::TYPE>::module[] = #MODULE;  \
   extern template struct node_declaration_traits<node::TYPE>;              \
   extern template struct node_definition_traits<node::TYPE, modules::_std::tag>
+
+#define YAVE_DEF_PRIM_NODE(TYPE)                       \
+  template struct node_declaration_traits<node::TYPE>; \
+  template struct node_definition_traits<node::TYPE, modules::_std::tag>
 
 namespace yave {
 
@@ -33,20 +39,13 @@ namespace yave {
   {
     static auto get_node_declaration() -> node_declaration;
     static const char node_name[];
+    static const char module[];
   };
 
   template <class T>
-  struct node_definition_traits<
-    node::PrimitiveCtor<T>,
-    modules::_std::tag>
+  struct node_definition_traits<node::PrimitiveCtor<T>, modules::_std::tag>
   {
     static auto get_node_definitions() -> std::vector<node_definition>;
   };
-
-  // Also need to add definitions in .cpp to property link with clang.
-  YAVE_DECL_PRIM_NODE(Int);
-  YAVE_DECL_PRIM_NODE(Float);
-  YAVE_DECL_PRIM_NODE(Bool);
-  YAVE_DECL_PRIM_NODE(String);
 
 } // namespace yave
