+++ src/yave/module/std/bool/bool.cpp
@@ -1,0 +1,9 @@
+//
+// Copyright (c) 2019 mocabe (https://github.com/mocabe)
+// Distributed under LGPLv3 License. See LICENSE for more details.
+//
+
+#include <yave/module/std/bool/bool.hpp>
+
+namespace yave {
+}
